--- include/ServerHeader.h
@@ -1,3 +1,0 @@
-#include "BlynkHeader.h"
-#include "ServerHeader.h"
-#include "HomekitHeader.h"
